// src/shared/api/generated/types/activityDtoType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Gateway.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { FileDtoType } from "./fileDtoType";
import type { ContactDtoType } from "./contactDtoType";

export interface ActivityDtoType {
  id?: string;
  description: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  maxParticipants?: number | null;
  startDate?: string;
  endDate?: string;
  /** @nullable */
  minAge?: number | null;
  /** @nullable */
  maxAge?: number | null;
  organizerId: string;
  price?: number;
  title: string;
  categoryIds?: number[];
  image?: FileDtoType;
  contacts?: ContactDtoType[];
}